Allow disabling the popper run timeout

+++ cli/popper/commands/cmd_run.py
@@ -25,7 +25,8 @@ from collections import defaultdict
     help='Timeout limit for pipeline. Use s for seconds, m for minutes and h '
          'for hours. A single integer can also be used to specify timeout '
          'in seconds. Use double quotes if you wish to use more than one unit.'
-         'For example: --timeout "2m 20s" will mean 140 seconds.',
+         'For example: --timeout "2m 20s" will mean 140 seconds. A value of 0'
+         'means no timeout',
     required=False,
     show_default=True,
     default="10800s"
@@ -410,7 +411,7 @@ def execute(stage, timeout, output_dir, bar=None):
 
         while p.poll() is None:
 
-            if time.time() > time_limit:
+            if timeout != 0.0 and time.time() > time_limit:
                 os.killpg(os.getpgid(p.pid), signal.SIGTERM)
                 sys.stdout.write(' time out!')
                 break
